requirements.txt
PySide6
ultralytics
paddleocr
paddlepaddle
dxcam; sys_platform == 'win32'
mss
pyautogui
pyyaml
loguru
litellm
openai
